Clang6

See merge request motis/motis!90

+++ base/core/include/motis/core/schedule/constant_graph.h
@@ -93,7 +93,7 @@ inline bool is_connected(node const* from, node const* to) {
 inline constant_graph build_interchange_graph(
     std::vector<station_node_ptr> const& station_nodes, unsigned route_count,
     search_dir const dir) {
-  auto const route_offset = station_nodes.size();
+  auto const route_offset = static_cast<uint32_t>(station_nodes.size());
   constant_graph g(route_offset + route_count);
 
   auto is_new = [&g, dir](uint32_t from, uint32_t to) {
